Refactor transaction input data resolvers and update import paths
- Introduced a new `resolver.ts` file to define the `TxInputDataResolver` type for improved type safety.
- Moved transaction input data resolver functions for various chains (Cardano, Cosmos, EVM, Polkadot, Ripple, Solana, Sui, Ton, Tron, UTXO) into a dedicated `resolvers` directory, enhancing organization and maintainability.
- Updated import paths across the codebase to reflect the new structure, ensuring consistency.
