+++ core/mpc/keysign/txInputData/resolvers/index.ts
@@ -2,18 +2,18 @@ import { ChainKind, getChainKind } from '@core/chain/ChainKind'
 import { KeysignPayload } from '@core/mpc/types/vultisig/keysign/v1/keysign_message_pb'
 import { WalletCore } from '@trustwallet/wallet-core'
 
-import { getKeysignChain } from '../utils/getKeysignChain'
-import { TxInputDataResolver } from './resolver'
-import { getCardanoTxInputData } from './resolvers/cardano'
-import { getCosmosTxInputData } from './resolvers/cosmos'
-import { getEvmTxInputData } from './resolvers/evm'
-import { getPolkadotTxInputData } from './resolvers/polkadot'
-import { getRippleTxInputData } from './resolvers/ripple'
-import { getSolanaTxInputData } from './resolvers/solana'
-import { getSuiTxInputData } from './resolvers/sui'
-import { getTonTxInputData } from './resolvers/ton'
-import { getTronTxInputData } from './resolvers/tron'
-import { getUtxoTxInputData } from './resolvers/utxo'
+import { getKeysignChain } from '../../utils/getKeysignChain'
+import { TxInputDataResolver } from '../resolver'
+import { getCardanoTxInputData } from './cardano'
+import { getCosmosTxInputData } from './cosmos'
+import { getEvmTxInputData } from './evm'
+import { getPolkadotTxInputData } from './polkadot'
+import { getRippleTxInputData } from './ripple'
+import { getSolanaTxInputData } from './solana'
+import { getSuiTxInputData } from './sui'
+import { getTonTxInputData } from './ton'
+import { getTronTxInputData } from './tron'
+import { getUtxoTxInputData } from './utxo'
 
 type Input = {
   keysignPayload: KeysignPayload
