+++ core/mpc/keysign/txInputData/resolvers/ripple.ts
@@ -2,9 +2,9 @@ import { assertField } from '@lib/utils/record/assertField'
 import { TW } from '@trustwallet/wallet-core'
 import Long from 'long'
 
-import { getBlockchainSpecificValue } from '../chainSpecific/KeysignChainSpecific'
-import { getKeysignTwPublicKey } from '../tw/getKeysignTwPublicKey'
-import { TxInputDataResolver } from './TxInputDataResolver'
+import { getBlockchainSpecificValue } from '../../chainSpecific/KeysignChainSpecific'
+import { getKeysignTwPublicKey } from '../../tw/getKeysignTwPublicKey'
+import { TxInputDataResolver } from '../resolver'
 
 export const getRippleTxInputData: TxInputDataResolver<'ripple'> = ({
   keysignPayload,
